+++ src/spelling/hunspellprovider.cpp
@@ -1,8 +1,7 @@
 /***********************************************************************
  *
- * Copyright (C) 2022 wereturtle
  * Copyright (C) 2009, 2010, 2011, 2012, 2013 Graeme Gott <graeme@gottcode.org>
- * Copyright (C) 2014-2020 wereturtle
+ * Copyright (C) 2014-2022 wereturtle
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by